Refactor worker and shared
