+++ distributed/dashboard/tests/test_scheduler_bokeh.py
@@ -16,7 +16,8 @@ from distributed.utils import tokey, format_dashboard_link
 from distributed.client import wait
 from distributed.metrics import time
 from distributed.utils_test import gen_cluster, inc, dec, slowinc, div, get_cert
-from distributed.dashboard.worker import Counters, BokehWorker
+from distributed.dashboard.worker import BokehWorker
+from distributed.dashboard.components.worker import Counters
 from distributed.dashboard.scheduler import applications, BokehScheduler
 from distributed.dashboard.components.scheduler import (
     SystemMonitor,
@@ -24,7 +25,7 @@ from distributed.dashboard.components.scheduler import (
     StealingTimeSeries,
     StealingEvents,
     Events,
-    TaskStream,
+    SchedulerTaskStream,
     TaskProgress,
     MemoryUse,
     CurrentLoad,
@@ -130,7 +131,7 @@ def test_events(c, s, a, b):
 
 @gen_cluster(client=True)
 def test_task_stream(c, s, a, b):
-    ts = TaskStream(s)
+    ts = SchedulerTaskStream(s)
 
     futures = c.map(slowinc, range(10), delay=0.001)
 
@@ -156,7 +157,7 @@ def test_task_stream(c, s, a, b):
 
 @gen_cluster(client=True)
 def test_task_stream_n_rectangles(c, s, a, b):
-    ts = TaskStream(s, n_rectangles=10)
+    ts = SchedulerTaskStream(s, n_rectangles=10)
     futures = c.map(slowinc, range(10), delay=0.001)
     yield wait(futures)
     ts.update()
@@ -166,19 +167,19 @@ def test_task_stream_n_rectangles(c, s, a, b):
 
 @gen_cluster(client=True)
 def test_task_stream_second_plugin(c, s, a, b):
-    ts = TaskStream(s, n_rectangles=10, clear_interval=10)
+    ts = SchedulerTaskStream(s, n_rectangles=10, clear_interval=10)
     ts.update()
     futures = c.map(inc, range(10))
     yield wait(futures)
     ts.update()
 
-    ts2 = TaskStream(s, n_rectangles=5, clear_interval=10)
+    ts2 = SchedulerTaskStream(s, n_rectangles=5, clear_interval=10)
     ts2.update()
 
 
 @gen_cluster(client=True)
 def test_task_stream_clear_interval(c, s, a, b):
-    ts = TaskStream(s, clear_interval=200)
+    ts = SchedulerTaskStream(s, clear_interval=200)
 
     yield wait(c.map(inc, range(10)))
     ts.update()
@@ -625,7 +626,7 @@ def test_proxy_to_workers(c, s, a, b):
 async def test_lots_of_tasks(c, s, a, b):
     import toolz
 
-    ts = TaskStream(s)
+    ts = SchedulerTaskStream(s)
     ts.update()
     futures = c.map(toolz.identity, range(100))
     await wait(futures)
